added check ica page
