+++ src/util/routes.tsx
@@ -7,6 +7,7 @@ import NotFound from "../pages/NotFound";
 import CreditorInstitution from "../pages/creditorInstitutions/CreditorInstitution";
 import BrokerPage from "../pages/brokers/BrokerPage";
 import Ica from "../pages/icas/ICAs";
+import CheckICA from "../pages/icas/CheckICA";
 
 export default class Routes extends React.Component {
 
@@ -23,6 +24,7 @@ export default class Routes extends React.Component {
                                     <Route path="/brokers" exact component={BrokersPage}/>
                                     <Route path="/brokers/:code" component={BrokerPage} />
                                     <Route path="/icas" exact component={Ica} />
+                                    <Route path="/icas/check" exact component={CheckICA} />
                                     <Route component={NotFound}/>
                                 </Switch>
                             </Layout>
